update final 13

+++ src/engine/vertex.rs
@@ -19,7 +19,7 @@ impl Vertex {
         Self { pos: pos }
     }
 
-    pub fn init_screen_space_transform(half_width: f64, half_height: f64) -> Matrix4 {
+    pub fn init_screen_space_transform(half_width: f32, half_height: f32) -> Matrix4 {
         Matrix4::init_screen_space_transform(half_width, half_height)
     }
 
@@ -35,7 +35,7 @@ impl Vertex {
         Vertex::new(transform.transform(self.pos))
     }
 
-    pub fn triangle_area_times_two(&self, b: Vertex, c: Vertex) -> f64 {
+    pub fn triangle_area_times_two(&self, b: Vertex, c: Vertex) -> f32 {
         let x1 = b.get_x() - self.pos.x;
         let y1 = b.get_y() - self.pos.y;
 
@@ -45,16 +45,16 @@ impl Vertex {
         x1 * y2 - x2 * y1
     }
 
-    pub fn get_x(self) -> f64 {
+    pub fn get_x(self) -> f32 {
         self.pos.x
     }
-    pub fn get_y(self) -> f64 {
+    pub fn get_y(self) -> f32 {
         self.pos.y
     }
-    pub fn get_z(self) -> f64 {
+    pub fn get_z(self) -> f32 {
         self.pos.z
     }
-    pub fn get_w(self) -> f64 {
+    pub fn get_w(self) -> f32 {
         self.pos.w
     }
